backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
cachecontrol==0.14.4
    # via firebase-admin
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
cryptography==46.0.4
    # via
    #   google-auth
    #   pyjwt
distro==1.9.0
    # via google-genai
fastapi==0.128.0
    # via -r requirements.in
firebase-admin==7.1.0
    # via -r requirements.in
google-api-core[grpc]==2.29.0
    # via
    #   firebase-admin
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth[requests]==2.48.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
    #   google-genai
google-cloud-core==2.5.0
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.23.0
    # via firebase-admin
google-cloud-storage==3.9.0
    # via firebase-admin
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.61.0
    # via -r requirements.in
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx[http2]==0.28.1
    # via
    #   firebase-admin
    #   google-genai
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
msgpack==1.1.2
    # via cachecontrol
proto-plus==1.27.1
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==6.33.5
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   google-genai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pyjwt[crypto]==2.11.0
    # via firebase-admin
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.22
    # via -r requirements.in
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via
    #   cachecontrol
    #   google-api-core
    #   google-auth
    #   google-cloud-storage
    #   google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via google-genai
starlette==0.50.0
    # via fastapi
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   google-genai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn[standard]==0.40.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
